+++ vips/color.c
@@ -1,4 +1,4 @@
-
+#include "color.h"
 
 int is_colorspace_supported(VipsImage *in) {
 	return vips_colourspace_issupported(in) ? 1 : 0;
