+++ vips/lang.h
@@ -1,0 +1,4 @@
+#include <stdlib.h>
+#include <vips/vips.h>
+
+#define INT_TO_GBOOLEAN(bool) (bool > 0 ? TRUE : FALSE)
